requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.5
django-cors-headers==3.1.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.1
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
